requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
Django==3.0.3
djangorestframework==3.11.0
flake8==3.8.2
flake8-polyfill==1.0.2
flake8-quotes @ file:///home/conda/feedstock_root/build_artifacts/flake8-quotes_1589330264223/work
importlib-metadata==1.6.0
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.3.0
numpy==1.18.1
olefile==0.46
opencv-contrib-python==4.2.0.34
packaging==20.3
pandas==1.0.4
pdf2image @ file:///home/conda/feedstock_root/build_artifacts/pdf2image_1588278609827/work
pep8-naming==0.8.2
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyparsing==2.4.7
pytesseract==0.3.4
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
torch==1.3.1
torchvision==0.4.2
wcwidth==0.1.9
zipp==3.1.0
